ansible: don't restart iptables service
There's no need to restart the iptables service here. The `iptables -I`
command takes effect right away, and the following `iptables save`
ensures that the change is permanent.

This patch also solves the issues described here:
https://github.com/kubernetes/kubernetes/pull/12601

(Unnecessarily restarting iptables causes the docker chain to be lost,
which in turn would have to be restarted, which in turn would require
the kubelet service to be restarted.)