ansible/roles/flannel: always try to restart docker
We don't want to put a conditional on the flannel task itself.
Otherwise, flannel will only be restarted on the nodes.

Additionally, we don't want to restart docker only on the nodes, but on
every host on which flannel and docker is installed. To do this, we
rely on the fact that if docker is running, then it means that the
docker0 interface will exist. Checking this is easy since the setup
module that is run at startup aggregates the available interfaces in the
ansible_interfaces variable.