Add the ability to specify a custom hyperkube base URL
While using the docker-multinode deployment strategy it became evident
that it would be beneficial for us to periodically roll our own
hyperkube images with a few extras installed to make things like NFS
work properly for our kubelets. Therefore, this commit introduces the
new HYPERKUBE_BASE_URL setting that allows someone invoking the script
to specify a custom base URL for a hyperkube image.

Arch and Kubernetes version will still be added onto the end to form the
final URL for the image and tag.