Fix API generation
Previous work to develop a Makefile that would handle API autogenerated
code was less than ideal. This change backs those changes out and solves
this in a more supportable way.

This change adds a one-time Dockerfile that will build an image with the
necessary api generation dependencies and runs against the local source
tree.

In addition, this removes the dependency on coversion-gen source.

Signed-off-by: Craig Tracey <craigtracey@gmail.com>