Increase default IO timeout submitted to NVMe devices
If there's a delay for EBS to respond, we can reach the timeout.
When this happens, the disk is re-mounted as read-only.
The default is 30 seconds.
AWS recommends to set it as high as possible. The newest Amazon Linux
has this configured too.
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html#timeout-nvme-ebs-volumes
For the kernel version, the highest value is 255.

Related issue: https://github.com/kubernetes/kops/issues/5694